Start to migrate DefaultCliAdapter tests for use with PicoCliDelegate
Make minor changes to the pico implementation code in response to test migration
