+++ cli/pico-cli/src/main/java/com/quorum/tessera/picocli/KeyGenCommand.java
@@ -20,84 +20,83 @@ import java.util.Set;
 import java.util.concurrent.Callable;
 
 @CommandLine.Command(
-    name = "keygen",
-    aliases = {"-keygen"},
-    headerHeading = "Usage:%n%n",
-    synopsisHeading = "%n",
-    descriptionHeading = "%nDescription:%n%n",
-    parameterListHeading = "%nParameters:%n",
-    optionListHeading = "%nOptions:%n",
-    header = "Generate Tessera encryption keys",
-    abbreviateSynopsis = true,
-    subcommands = {CommandLine.HelpCommand.class}
-)
+        name = "keygen",
+        aliases = {"-keygen"},
+        headerHeading = "Usage:%n%n",
+        synopsisHeading = "%n",
+        descriptionHeading = "%nDescription:%n%n",
+        parameterListHeading = "%nParameters:%n",
+        optionListHeading = "%nOptions:%n",
+        header = "Generate Tessera encryption keys",
+        abbreviateSynopsis = true,
+        subcommands = {CommandLine.HelpCommand.class})
 public class KeyGenCommand implements Callable<CliResult> {
     private final KeyGeneratorFactory factory = KeyGeneratorFactory.newFactory();
 
     private final Validator validator =
-        Validation.byDefaultProvider().configure().ignoreXmlConfiguration().buildValidatorFactory().getValidator();
+            Validation.byDefaultProvider().configure().ignoreXmlConfiguration().buildValidatorFactory().getValidator();
 
     // TODO(cjh) raise CLI usage wording changes as separate change
 
     @CommandLine.Option(
-        names = {"--output", "-filename"},
-        description = "Comma-separated list of paths to save generated key files. Can also be used with keyvault. Number of args determines number of key-pairs generated."
-    )
-    public List<String> output;
+            names = {"--keyout", "-filename"},
+            defaultValue = ".",
+            description =
+                    "Comma-separated list of paths to save generated key files. Can also be used with keyvault. Number of args determines number of key-pairs generated (default = ${DEFAULT-VALUE})")
+    public List<String> keyOut;
 
     // TODO(cjh) review description and name
     @CommandLine.Option(
-        names = {"--encryptionconfig", "-keygenconfig"},
-        description = "File containing Argon2 encryption config used to secure the new private key"
-    )
+            names = {"--encryptionconfig", "-keygenconfig"},
+            description = "File containing Argon2 encryption config used to secure the new private key")
     public ArgonOptions encryptionConfig;
 
     @CommandLine.Option(
-        names = {"--vault.type", "-keygenvaulttype"},
-        description = "Specify the key vault provider the generated key is to be saved in.  If not set, the key will be encrypted and stored on the local filesystem.  Valid values: ${COMPLETION-CANDIDATES})"
-    )
-    //TODO(cjh) get possible enum values to show in the usage
+            names = {"--vault.type", "-keygenvaulttype"},
+            description =
+                    "Specify the key vault provider the generated key is to be saved in.  If not set, the key will be encrypted and stored on the local filesystem.  Valid values: ${COMPLETION-CANDIDATES})")
+    // TODO(cjh) get possible enum values to show in the usage
     public KeyVaultType vaultType;
 
     @CommandLine.Option(
-        names = {"--vault.url", "-keygenvaulturl"},
-        description = "Base url for key vault"
-    )
+            names = {"--vault.url", "-keygenvaulturl"},
+            description = "Base url for key vault")
     public String vaultUrl;
 
     @CommandLine.Option(
-        names = {"--vault.hashicorp.approlepath", "-keygenvaultapprole"},
-        description = "AppRole path for Hashicorp Vault authentication (defaults to 'approle')"
-    )
+            names = {"--vault.hashicorp.approlepath", "-keygenvaultapprole"},
+            description = "AppRole path for Hashicorp Vault authentication (defaults to 'approle')")
     public String hashicorpApprolePath;
 
     @CommandLine.Option(
-        names = {"--vault.hashicorp.secretenginepath", "-keygenvaultsecretengine"},
-        description = "Name of already enabled Hashicorp v2 kv secret engine"
-    )
+            names = {"--vault.hashicorp.secretenginepath", "-keygenvaultsecretengine"},
+            description = "Name of already enabled Hashicorp v2 kv secret engine")
     public String hashicorpSecretEnginePath;
 
     @CommandLine.Option(
-        names = {"--vault.hashicorp.tlskeystore", "-keygenvaultkeystore"},
-        description = "Path to JKS keystore for TLS Hashicorp Vault communication"
-    )
+            names = {"--vault.hashicorp.tlskeystore", "-keygenvaultkeystore"},
+            description = "Path to JKS keystore for TLS Hashicorp Vault communication")
     public Path hashicorpTlsKeystore;
 
     @CommandLine.Option(
-        names = {"--vault.hashicorp.tlstruststore", "-keygenvaulttruststore"},
-        description = "Path to JKS truststore for TLS Hashicorp Vault communication"
-    )
+            names = {"--vault.hashicorp.tlstruststore", "-keygenvaulttruststore"},
+            description = "Path to JKS truststore for TLS Hashicorp Vault communication")
     public Path hashicorpTlsTruststore;
 
-    //TODO(cjh) do something about the duplication of the configfile option in each relevant command
+    // TODO(cjh) do something about the duplication of the configfile option in each relevant command
     @CommandLine.Option(
-        names = {"--configfile", "-configfile"},
-        description = "Path to node configuration file"
-    )
+            names = {"--configfile", "-configfile"},
+            description = "Path to node configuration file")
     public Config config;
 
-    @CommandLine.Mixin
-    public EncryptorOptions encryptorOptions;
+    // TODO(cjh) implement config output
+    @CommandLine.Option(
+        names = {"--configout", "-output"},
+        description =
+            "Path to save updated configfile to.  Updated config will be printed to terminal if not provided.  Only valid if --configfile option also provided.")
+    public List<String> configOut;
+
+    @CommandLine.Mixin public EncryptorOptions encryptorOptions;
 
     @Override
     public CliResult call() throws Exception {
@@ -114,9 +113,7 @@ public class KeyGenCommand implements Callable<CliResult> {
 
         final KeyGenerator generator = factory.create(keyVaultConfig, encryptorConfig);
 
-        output.forEach(
-            name -> generator.generate(name, encryptionConfig, keyVaultOptions)
-        );
+        keyOut.forEach(name -> generator.generate(name, encryptionConfig, keyVaultOptions));
 
         return new CliResult(0, true, null);
     }
@@ -140,23 +137,23 @@ public class KeyGenCommand implements Callable<CliResult> {
             keyVaultConfig = new AzureKeyVaultConfig(vaultUrl);
 
             Set<ConstraintViolation<AzureKeyVaultConfig>> violations =
-                validator.validate((AzureKeyVaultConfig) keyVaultConfig);
+                    validator.validate((AzureKeyVaultConfig) keyVaultConfig);
 
             if (!violations.isEmpty()) {
                 throw new ConstraintViolationException(violations);
             }
         } else {
-            if (output.size() == 0) {
+            if (keyOut.size() == 0) {
                 throw new CliException(
-                    "At least one -filename must be provided when saving generated keys in a Hashicorp Vault");
+                        "At least one -filename must be provided when saving generated keys in a Hashicorp Vault");
             }
 
             keyVaultConfig =
-                new HashicorpKeyVaultConfig(
-                    vaultUrl, hashicorpApprolePath, hashicorpTlsKeystore, hashicorpTlsTruststore);
+                    new HashicorpKeyVaultConfig(
+                            vaultUrl, hashicorpApprolePath, hashicorpTlsKeystore, hashicorpTlsTruststore);
 
             Set<ConstraintViolation<HashicorpKeyVaultConfig>> violations =
-                validator.validate((HashicorpKeyVaultConfig) keyVaultConfig);
+                    validator.validate((HashicorpKeyVaultConfig) keyVaultConfig);
 
             if (!violations.isEmpty()) {
                 throw new ConstraintViolationException(violations);
