+++ cli/pico-cli/src/main/java/com/quorum/tessera/picocli/NoTesseraConfigfileOptionException.java
@@ -1,0 +1,5 @@
+package com.quorum.tessera.picocli;
+
+public class NoTesseraConfigfileOptionException extends RuntimeException {
+
+}
