+++ cli/pico-cli/src/test/resources/META-INF/services/com.quorum.tessera.key.generation.KeyGeneratorFactory
@@ -1,0 +1,1 @@
+com.quorum.tessera.picocli.keys.MockKeyGeneratorFactory
