+++ cli/pico-cli/src/test/resources/lockedprivatekey.json
@@ -1,0 +1,14 @@
+{
+    "data": {
+        "aopts": {
+            "variant": "id",
+            "memory": 1024,
+            "iterations": 1,
+            "parallelism": 1
+        },
+        "snonce": "dwixVoY+pOI2FMuu4k0jLqN/naQiTzWe",
+        "asalt": "JoPVq9G6NdOb+Ugv+HnUeA==",
+        "sbox": "6Jd/MXn29fk6jcrFYGPb75l7sDJae06I3Y1Op+bZSZqlYXsMpa/8lLE29H0sX3yw"
+    },
+    "type": "argon2sbox"
+}
